feat: add functionality to unassign and remove officers from reports in report services #182

+++ participium/src/app/lib/services/user.service.ts
@@ -106,7 +106,11 @@ class UserService {
     if (!officer || !officer.role.includes("TECHNICAL_OFFICER")) {
       throw new Error(`Officer with ID ${userId} not found`);
     }
-    await ReportAssignmentService.getInstance().unassignReportsOfDeletedOfficer(officer.managedReports);
+    const unassigned = await ReportAssignmentService.getInstance().unassignReportsOfDeletedOfficer(officer.managedReports);
+    
+    if (!unassigned) {
+      return false;
+    }
 
     return await this.userRepository.deleteOfficer(userId);
   }
